chore: refactor files structure
