+++ src/Apps/Consign/Routes/SubmissionFlow/SoldRecentlyOnArtsy.tsx
@@ -177,6 +177,7 @@ const PLACEHOLDER = (
 )
 
 export const SoldRecentlyOnArtsyQueryRenderer: React.FC = () => {
+  console.log("[LOGD] HELLO ")
   return (
     <SystemQueryRenderer<SoldRecentlyOnArtsyQuery>
       lazyLoad
