+++ src/Apps/Consign/consignFromCollectorProfileMyCollectionRoutes.tsx
@@ -18,7 +18,7 @@ const MarketingLandingApp = loadable(
 const FAQApp = loadable(
   () =>
     import(
-      /* webpackChunkName: "consignBundle" */ "./Routes/MarketingLanding/FAQApp"
+      /* webpackChunkName: "consignBundle" */ "./Routes/SubmissionFlow/FAQ/FAQApp"
     ),
   {
     resolveComponent: component => component.FAQApp,
